challenges/Module A - Inside-Out (classic-stateless)/3 - Manage Complexity (TPP)/3.2 - Boolean Calculator/src/index.spec.ts
import { evaluate } from "./index";

describe('boolean calculator', () => {
    describe('single statements', () => {
        it('TRUE should return true', () => {
            const input = 'TRUE';

            expect(evaluate(input)).toBeTruthy();
        })

        it('FALSE should return false', () => {
            const input = 'FALSE';

            expect(evaluate(input)).toBeFalsy();
        })
    })

    describe('simple combinations', () => {
        it('TRUE AND TRUE should return true', () => {
            const input = 'TRUE AND TRUE';

            expect(evaluate(input)).toBeTruthy();
        })
    })
})
